Remove test scripts
