+++ package.json
@@ -67,9 +67,7 @@
       "bella": "https://raw.githubusercontent.com/techpush/bella.js/master/src/bella.js",
       "fetch": "https://raw.githubusercontent.com/typicode/fetchival/master/index.js",
       "promise": "https://raw.githubusercontent.com/jakearchibald/es6-promise/master/dist/es6-promise.js",
-      "sinon": "http://sinonjs.org/releases/sinon-1.17.3.js",
-      "t3native": "https://raw.githubusercontent.com/box/t3js/master/dist/t3-native-2.4.0.js",
-      "t3nativeTest": "https://raw.githubusercontent.com/box/t3js/master/dist/t3-native-testing.js"
+      "t3native": "https://raw.githubusercontent.com/box/t3js/master/dist/t3-native-2.4.0.js"
     }
   }
 }
