+++ src/window.html
@@ -53,7 +53,7 @@
   <script type="text/javascript" src="/js/vendor/fetch.js"></script>
   <script type="text/javascript" src="/js/vendor/promise.js"></script>
   <script type="text/javascript" src="/js/services/wordlist.js"></script>
-  <script type="text/javascript" src="/js/services/sentence.js"></script>
+  <script type="text/javascript" src="/js/services/generator.js"></script>
   <script type="text/javascript" src="/js/services/storage.js"></script>
   <script type="text/javascript" src="/js/services/parser.js"></script>
 <script type="text/javascript" src="/js/modules/summary.js"></script>
